Fix utf8 encoding issues
Google's login form includes a hidden field called _utf8 which contains \u2603,
which is a snowman, probably as a workaround to force IE5-8 to read and encode
things in UTF-8.

This doesn't interact very well with urllib.urlencode, which bails out with a
ValueError, so this commit pre-encodes strings using .encode() to avoid this
scenario.